Get tests up and running
