+++ __tests__/net-server-lists.test.ts
@@ -1,6 +1,5 @@
-import ElectrumClient = require('electrum-client');
-
 jest.mock('electrum-client');
+import ElectrumClient from 'electrum-client';
 
 describe('network functions with bad server lists', () => {
     beforeEach(() => {
